requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiomysql==0.2.0
aiosignal==1.4.0
alembic==1.16.5
async-timeout==5.0.1
attrs==25.3.0
backports.asyncio.runner==1.2.0
cffi==2.0.0
cryptography==45.0.7
exceptiongroup==1.3.0
frozenlist==1.7.0
greenlet==3.2.4
idna==3.10
iniconfig==2.1.0
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.6.4
packaging==25.0
pluggy==1.6.0
propcache==0.4.0
pycparser==2.23
Pygments==2.19.2
PyMySQL==1.1.2
pytest==8.4.2
pytest-asyncio==1.2.0
PyYAML==6.0.2
SQLAlchemy==2.0.43
tomli==2.2.1
typing_extensions==4.15.0
websocat==1.13.0
websockets==15.0.1
yarl==1.20.1